Refactor node order plugin
