+++ pkg/scheduler/plugins/nodeorder/nodeaffinity.go
@@ -1,0 +1,32 @@
+/*
+Copyright 2022 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package nodeorder
+
+import (
+	v1 "k8s.io/api/core/v1"
+	"k8s.io/kubernetes/pkg/scheduler/apis/config"
+	k8sframework "k8s.io/kubernetes/pkg/scheduler/framework"
+	"k8s.io/kubernetes/pkg/scheduler/framework/plugins/nodeaffinity"
+)
+
+func nodeAffinity(handle k8sframework.Handle) (*nodeaffinity.NodeAffinity, error) {
+	naArgs := &config.NodeAffinityArgs{
+		AddedAffinity: &v1.NodeAffinity{},
+	}
+	p, err := nodeaffinity.New(naArgs, handle)
+	return p.(*nodeaffinity.NodeAffinity), err
+}
